Unified Remove Function
The new `remove` function unifies both `deleteFile` and `deleteFolder`.
(It can't be called `delete`, which is already reserved by the low level API)

It determines whether the source is a file or folder from the server response
(HEAD call is reused in case of file deletion) instead of the presence of a trailing slash on provided source url.